Create README.md
“Body Hero” is a mobile based web application for IQRAA COMMUNITY CLINIC,
THOTTILPALAM.Skin is the largest organ in human body, which is important to cover
human bone, and to protect human from any harm, fight the bacteria and other kind of
diseases, and may have numerous potential abnormalities. Several factors may affect the
skin directly or indirectly and cause diseases which can be treated with specific medicine
and others require doctor’s consultation. Skin disease is the most common disease in the
world. The diagnosis of the skin disease requires a high level of expertise and accuracy
for dermatologist, so computer aided skin disease diagnosis model is proposed to provide
more objective and reliable solution. Many researches were done to detect skin diseases
like skin cancer and tumor skin. But the accurate recognition of the disease is extremely
challenging due to the many reasons.
There are many skin diseases including skin cancer, eczema, allergies. The focus of this
project will be the prediction of different types of skin related diseases, where the image
will be taken either directly from the mobile camera or choose it from database and then
analyzes the image and isolate the skin area and extract the affected area.
As for the technical accuracy, the application will implement the prediction system using
the convolutional neural network which is the highly trusted methodology for prediction
systems.